Setup router password with dashboard auth
We don't want to lock users out of their own routers, so we've been
shipping with password auth enabled on the routers but not default password
set, this is obviously a huge security hole. This piggybacks setting the ssh
password to setting the router dashboard password, so that users will be able
to access their devices but they won't be left insecure.